Move calculator to flow directory - landlord-immigration-check
This calculator is specific to the landlord-immigration-check flow and so should
be co-located.

* SmartAnswer::Calculators:LandlordImmigrationCheckCalculator
  -> SmartAnswer::LandlordImmigrationCheckFlow::Calculator

* SmartAnswer::Calculators::LandlordImmigrationCheckCalculatorTest
  -> SmartAnswer::LandlordImmigrationCheckFlow::CalculatorTest

Since these changes do not affect the behaviour and all the regression tests
pass, I've updated the checksums with the new location of the calculator.
